feat: add service for transactions
- Creates a transactionsService.js file

- Adds a function to fetch transactions from the API

- Centralizes the logic for transaction requests

- Improves code maintainability and reusability
